rebase botmessage model
