+++ backend/content/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 
 from content.models import (
-    BotMessage,
     Category,
     ContentFile,
     ContentRating,
@@ -84,12 +83,6 @@ class ContentRatingAdmin(admin.ModelAdmin):
 
     def get_queryset(self, request):
         return super().get_queryset(request).select_related('content', 'user')
-
-
-@admin.register(BotMessage)
-class BotMessageAdmin(admin.ModelAdmin):
-    list_display = ['key', 'text', 'updated_at']
-    search_fields = ['key', 'text', 'comment']
 
 
 @admin.register(ContentViewStat)
