+++ backend/content/models.py
@@ -159,33 +159,6 @@ class ContentRating(models.Model):
         return Truncator(f'{self.user}\'s rating').chars(MAX_OBJECT_CHARS)
 
 
-class BotMessage(models.Model):
-    """Bot messages."""
-    key = models.CharField(
-        max_length=100,
-        unique=True,
-        verbose_name='message key'
-    )
-    text = models.TextField(verbose_name='message text')
-    comment = models.CharField(
-        max_length=255,
-        blank=True,
-        verbose_name='comment'
-    )
-    updated_at = models.DateTimeField(
-        auto_now=True,
-        verbose_name='Update time'
-    )
-
-    class Meta:
-        verbose_name = 'bot message'
-        verbose_name_plural = 'Bot messages'
-        ordering = ['key']
-
-    def __str__(self):
-        return self.key
-
-
 class ContentViewStat(models.Model):
     """User viewing statistics."""
